move code to correctly spelled interactions file
